+++ src/main/java/com/sbproject/piclocket/service/PhotoService.java
@@ -18,14 +18,23 @@ public class PhotoService {
 
     private static final Logger log = LoggerFactory.getLogger(PhotoService.class);
     private static final String DEMO_USER_ID = "demo-user-id";
+
     private final PhotoRepository photoRepository;
     private final S3PresignedUrlService s3PresignedUrlService;
+    private final S3ObjectService s3ObjectService;
 
-    public PhotoService(PhotoRepository photoRepository, S3PresignedUrlService s3PresignedUrlService) {
+    public PhotoService(PhotoRepository photoRepository, S3PresignedUrlService s3PresignedUrlService, S3ObjectService s3ObjectService) {
         this.photoRepository = photoRepository;
         this.s3PresignedUrlService = s3PresignedUrlService;
+        this.s3ObjectService = s3ObjectService;
     }
 
+    /**
+     * Fetches presigned URL and returns it to the client
+     *
+     * @param request contains photo metadata we will store in DB
+     * @return {@link CreateUploadResponse} contains the url for direct upload and photo id
+     */
     public CreateUploadResponse createUploadRequest(CreateUploadRequest request) {
         UUID photoId = UUID.randomUUID();
         Instant now = Instant.now();
@@ -58,5 +67,35 @@ public class PhotoService {
         );
 
         return new CreateUploadResponse(photoId, uploadUrl);
+    }
+
+    /**
+     * Checks whether or not a photo upload has successfully been completed and adjusts state accordingly
+     * If successful, state is updated to UPLOADED
+     *
+     * @param photoId the unique identifier of a photo
+     */
+    public void completeUpload(UUID photoId) {
+        Photo photo = photoRepository.findById(photoId)
+                .orElseThrow(() -> new IllegalArgumentException(
+                        "Photo not found for id=%s".formatted(photoId)
+                ));
+
+        boolean photoExists = s3ObjectService.objectExists(photo.getS3Key());
+
+        if (!photoExists) {
+            log.warn("Upload completion requested but S3 object was not found for photoId={}", photoId);
+
+            throw new IllegalStateException(
+                    "Photo upload has not completed for id=%s".formatted(photoId)
+            );
+        }
+
+        Instant now = Instant.now();
+
+        photo.markUploaded(now);
+        photoRepository.save(photo);
+
+        log.info("Upload completed for photoId={}, status={}", photoId, PhotoStatus.UPLOADED);
     }
 }